Task8

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,29 @@
 ﻿using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager) 
         {
             _context = context;
+            _userManager = userManager;
         }
+
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            ApplicationUser loggedInUser = _userManager.GetUserAsync(HttpContext.User);
+            var model = _context.Items.ToList().Where(s => s.User == loggedInUser);
 
             return View("Index", model);
         }
@@ -29,6 +37,8 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            var loggedInUser = _userManager.GetUserAsync(HttpContext.User);
+            item.User = loggedInUser;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -36,9 +46,17 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+            ApplicationUser loggedInUser = _userManager.GetUserAsync(HttpContext.User);
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
-            _context.SaveChanges();
+            if (item.User == loggedInUser)
+            {
+                _context.Items.Remove(item);
+                _context.SaveChanges();
+            }
+            else
+            {
+                return Unauthorized();
+            }
             return RedirectToAction("Index");
         }
     }
